split DocumentEditor/index into shared/ui
