+++ packages/fields-document/src/DocumentEditor/tests/utils.tsx
@@ -16,7 +16,8 @@ import React from 'react'
 import { act, render } from '@testing-library/react'
 import { diff } from 'jest-diff'
 import prettyFormat, { plugins, type Plugin } from 'pretty-format'
-import { createDocumentEditor, DocumentEditorEditable } from '..'
+import { DocumentEditorEditable } from '..'
+import { createDocumentEditor } from '../editor-shared'
 import { type ComponentBlock } from '../../component-blocks'
 import { type DocumentFeatures } from '../../views'
