+++ packages/fields-document/src/component-blocks.tsx
@@ -1,14 +1,18 @@
-export { NotEditable, component, fields } from './DocumentEditor/component-blocks/api'
+export {
+  NotEditable,
+  component,
+  fields
+} from './DocumentEditor/component-blocks/api'
 export type {
+  ArrayField,
   ChildField,
   ComponentBlock,
   ComponentSchema,
+  ComponentSchemaForGraphQL,
   ConditionalField,
   FormField,
+  InferRenderersForComponentBlocks,
   ObjectField,
   RelationshipData,
   RelationshipField,
-  InferRenderersForComponentBlocks,
-  ArrayField,
-  ComponentSchemaForGraphQL,
-} from './DocumentEditor/component-blocks/api'
+} from './DocumentEditor/component-blocks/api-shared'
