+++ packages/fields-document/src/validation.ts
@@ -1,10 +1,10 @@
 import { Text, Editor } from 'slate'
-import { createDocumentEditor } from './DocumentEditor'
 import {
   type ComponentBlock,
   type ComponentSchema
 } from './DocumentEditor/component-blocks/api-shared'
 import { assertNever, type ReadonlyPropPath } from './DocumentEditor/component-blocks/utils'
+import { createDocumentEditor } from './DocumentEditor/editor-shared'
 import { type Relationships } from './DocumentEditor/relationship'
 import {
   type ElementFromValidation,
