+++ packages/fields-document/src/views.tsx
@@ -15,7 +15,8 @@ import {
 } from '@keystone-6/core/types'
 import weakMemoize from '@emotion/weak-memoize'
 import { CellContainer, CellLink } from '@keystone-6/core/admin-ui/components'
-import { createDocumentEditor, DocumentEditor } from './DocumentEditor'
+import { createDocumentEditor } from './DocumentEditor/editor-shared'
+import { DocumentEditor } from './DocumentEditor'
 import { type ComponentBlock } from './component-blocks'
 import { type Relationships } from './DocumentEditor/relationship'
 import { clientSideValidateProp } from './DocumentEditor/component-blocks/utils'
